Update README.md
In the above code, three different sine waves with different frequency, amplitudes and phases are are generated on the same graph. Subplot was used to generate 25 plots of the same graph. The next code will attempt to randomly create missing values from 4% to 100% of the plotted data and then use imputation techniques to generate a code to predict the missing data